pyproject.toml
[project]
name = "ai-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.13",
]
